moved SEO component  in new folder -> adapted imports
