+++ src/pages/404.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 
 import Layout from "../components/Layout"
-import SEO from "../components/Seo"
+import SEO from "../components/Blog/SEO"
 
 const NotFoundPage: React.FC = () => (
   <Layout>
