+++ src/pages/about.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import Layout from "../components/Layout"
-import SEO from "../components/Seo"
+import SEO from "../components/Blog/SEO"
 
 const SecondPage: React.FC = () => (
   <Layout>
