+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import Layout from "../components/Layout"
-import SEO from "../components/Seo"
+import SEO from "../components/Blog/SEO"
 
 const IndexPage: React.FC = () => (
   <Layout showTeaser={true}>
